refactor(app): guideline annotation card
