+++ app/lib/common/pages/drug/widgets/module.dart
@@ -1,3 +1,4 @@
-export 'annotation_card.dart';
+export 'annotation_cards/drug.dart';
+export 'annotation_cards/guideline.dart';
 export 'disclaimer.dart';
 export 'sub_header.dart';
